fix: 게시글 조회, 댓글 조회 API mock data 수정
