+++ src/__mock__/data/comment.mock.ts
@@ -10,18 +10,20 @@ export const commentsMock = [
     created_at: '2023-12-01 23:06:10',
     updated_at: '2023-12-01 23:06:10',
     is_like: true,
-    children: [
+    deleted: 0,
+    childComments: [
       {
         id: 3,
-        content: '댓글입니다!!!!!!',
+        content: 'Comment Deleted',
         parent_id: 1,
-        user_id: 3,
-        user_nickname: '테스터닉네임',
-        user_profile_image: 'http://via.placeholder.com/360x360',
-        like_count: 21,
+        user_id: -9999,
+        user_nickname: 'default nickname',
+        user_profile_image: 'default profile image',
         created_at: '2023-12-01 23:06:10',
         updated_at: '2023-12-02 07:06:10',
+        like_count: 0,
         is_like: false,
+        deleted: 1,
       },
       {
         id: 4,
@@ -30,25 +32,27 @@ export const commentsMock = [
         user_id: 4,
         user_nickname: '댓글단사람',
         user_profile_image: 'http://via.placeholder.com/360x360',
-        like_count: 21,
         created_at: '2023-12-02 23:06:10',
         updated_at: '2023-12-02 23:06:10',
+        like_count: 23,
         is_like: true,
+        deleted: 0,
       },
     ],
   },
   {
     id: 2,
-    content: '원댓글입니다아아아앙',
+    content: 'Comment Deleted',
     parent_id: 0,
-    user_id: 8,
-    user_nickname: '늦게댓글',
-    user_profile_image: 'http://via.placeholder.com/360x360',
-    like_count: 21,
+    user_id: -9999,
+    user_nickname: 'default nickname',
+    user_profile_image: 'default profile image',
     created_at: '2023-12-05 11:16:10',
     updated_at: '2023-12-07 23:06:10',
-    is_like: true,
-    children: [
+    like_count: 0,
+    is_like: false,
+    deleted: 1,
+    childComments: [
       {
         id: 8,
         content: '대댓글띠',
@@ -56,10 +60,11 @@ export const commentsMock = [
         user_id: 4,
         user_nickname: '댓글단사람',
         user_profile_image: 'http://via.placeholder.com/360x360',
-        like_count: 21,
         created_at: '2023-12-07 23:06:10',
         updated_at: '2023-12-08 07:06:10',
+        like_count: 4,
         is_like: false,
+        deleted: 0,
       },
     ],
   },
