+++ src/__mock__/data/posts.mock.ts
@@ -1,0 +1,77 @@
+export const postsMock = [
+  {
+    id: 43,
+    title: '자유로운 게시글 제목55555',
+    content: '자유로운 게시글 본문~~~~~~~~~~55555555',
+    view_count: 76,
+    thumbnail_image: 'http://via.placeholder.com/360x360',
+    category: 1,
+    user_id: 3,
+    user_nickname: '테스터3',
+    user_profile_image: 'http://via.placeholder.com/360x360',
+    created_at: '2023-12-11 23:05:04',
+    updated_at: '2023-12-11 23:05:10',
+    like_count: 21,
+    comment_count: 13,
+  },
+  {
+    id: 42,
+    title: '자유로운 게시글 제목4444',
+    content: '자유로운 게시글 본문~~~~~~~~~~444444',
+    view_count: 82,
+    thumbnail_image: 'http://via.placeholder.com/360x360',
+    category: 1,
+    user_id: 4,
+    user_nickname: '테스터4',
+    user_profile_image: 'http://via.placeholder.com/360x360',
+    created_at: '2023-12-10 23:05:04',
+    updated_at: '2023-12-10 23:05:10',
+    like_count: 1,
+    comment_count: 0,
+  },
+  {
+    id: 40,
+    title: '자유로운 게시글 제목333',
+    content: '자유로운 게시글 본문~~~~~~~~~~3333333',
+    view_count: 133,
+    thumbnail_image: 'http://via.placeholder.com/360x360',
+    category: 1,
+    user_id: 2,
+    user_nickname: '테스터2',
+    user_profile_image: 'http://via.placeholder.com/360x360',
+    created_at: '2023-12-09 23:05:04',
+    updated_at: '2023-12-09 23:05:10',
+    like_count: 42,
+    comment_count: 32,
+  },
+  {
+    id: 35,
+    title: '자유로운 게시글 제목2',
+    content: '자유로운 게시글 본문~~~~~~~~~~222222',
+    view_count: 39,
+    thumbnail_image: 'http://via.placeholder.com/360x360',
+    category: 1,
+    user_id: 3,
+    user_nickname: '테스터3',
+    user_profile_image: 'http://via.placeholder.com/360x360',
+    created_at: '2023-12-06 23:05:04',
+    updated_at: '2023-12-06 23:05:10',
+    like_count: 9,
+    comment_count: 5,
+  },
+  {
+    id: 32,
+    title: '자유로운 게시글 제목1',
+    content: '자유로운 게시글 본문~~~~~~~~~~111111',
+    view_count: 195,
+    thumbnail_image: 'http://via.placeholder.com/360x360',
+    category: 1,
+    user_id: 1,
+    user_nickname: '테스터1',
+    user_profile_image: 'http://via.placeholder.com/360x360',
+    created_at: '2023-12-04 23:05:04',
+    updated_at: '2023-12-04 23:05:10',
+    like_count: 17,
+    comment_count: 22,
+  },
+];
