+++ src/constants/api/end-point.ts
@@ -1,4 +1,4 @@
 export const END_POINT = {
   detailPost: (id: number | string) => `/api/posts/${id}`,
-  post: '/api/posts',
+  posts: '/api/posts',
 };
